+++ Hardware/MotorThread.h
@@ -19,7 +19,7 @@ public:
 	MotorThread(const char* name);
 	void init();
 	void run();
-	void setMotorSpeed(uint8_t newMotorSpeed);
+	void setMotorSpeed(uint16_t newMotorSpeed);
 };
 
 #endif /* MOTORTHREAD_H_ */
