+++ PID/AnglePID.h
@@ -32,7 +32,7 @@ public:
 	AnglePID(void);
 	void setDestinationAngle(float angle);
 	void run(void);
-	float currentOutput(void);
+	uint16_t currentOutput(void);
 
 	void setP(float p);
 	void setI(float i);
